+++ src/gsfopt.cpp
@@ -110,7 +110,7 @@ std::string GsfOpt::ToTimeString(double t, bool padding)
 	if (!padding)
 	{
 		size_t len = strlen(str);
-		for (int i = len - 1; i >= 0; i--)
+		for (size_t i = len - 1; i >= 0; i--)
 		{
 			if (str[i] == '0')
 			{
